Renames adaptor to adapter. (#233907)
