+++ src/vs/editor/contrib/inlineCompletions/browser/model/suggestWidgetAdapter.ts
